+++ src/shape.c
@@ -2,7 +2,7 @@
 #include "shape.h"
 #include "debug.h"
 
-void shape_move(struct shape *shape, short delta_x, short delta_y) {
+void shape_move(struct shape *shape, int delta_x, int delta_y) {
     NULL_CHECK("shape_move()", shape);
     
     shape->offset[0] += delta_x;
